requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-templated-mail==1.1.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.2
djoser==1.5.1
Pillow==6.0.0
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
